requirements.txt
langchain_openai 
langchain_core
python-dotenv
streamlit
langchain_community
langserve
fastapi
uvicorn
sse_starlette
pypdf
chromadb
groq
cassio
beautifulsoup4
langchain-groq
langchainhub
ipykernel
langchain
pdfplumber